requirements.txt
streamlit
requests
pillow
opencv-python
pyzbar
numpy
verbwire
web3
